% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_package_content.R
\name{an_important_value}
\alias{an_important_value}
\title{Just some example function.
But this function will also be present in another package.}
\usage{
an_important_value()
}
\description{
Just some example function.
But this function will also be present in another package.
}
